agents/requirements.txt
# Insurance AI Demo - Agent Dependencies
openai>=1.40.0
azure-identity>=1.17.0
python-dotenv>=1.0.0
